+++ src/components/display/fewlines/Locale/Locale.tsx
